Created first App and Templates
Created an "Apps" directory to hold all the files associated with this project.  Then created the first App called "green".  I then added to urls and views appropriately to set up the first template, called "main".  Also created a base.html file in order to inherit from in the future.  Created a model called "Centers" to hold the data for recycling centers in this project.   Boolean fields were chosen based on recyclable items found on the Hall County recycling website.

I then changed settings in order to access and migrate data to the database.  I also added to installed apps for the future (bootstrap, etc).  Also added settings related to REST API which I will begin integrating next.

Finally I started the server, logged into the admin page, and created a "center" in order to test the framework I had set up.  I also added a small amount of html to "main" as a placeholder and to test the url.