Finished setting up API
I created a new file, api_urls and connected it to the project urls folder in order to set up the API urls.  I also edited the json_views in order to make the Filter work as well as make more sense.  I also changed the name to CenterCollection for clarity purposes.