Added first web map template
Created a new view called MapView.  Also added a new url to connect the new template, "center_map".  On the center_map page I added links to the api for MapBox, and added a prebuilt map template for Hall County.  Added links/script to base in order to connect to MapBox as well.