Created produce template
Created the produce template to display local produce stands on a map.  Created a view and url to sync up to the template.  Made the link in the nav bar, found in base html, work for the new produce template.  Also changed the text color to white in main to make it more viewable as the website gets set up.