Edited Stand model
Edited the fields of the stand model based on a how I plan on implementing adding the model to the database (based on a form registered users can access).  Changed around admin.py during an attempt at implementing another plan.  Changed it back to what it is now.  Also added a serializer in the first step of setting up the REST API.