Set up REST API
Finished setting up the REST API for produce stands model.  Created a new json_view as well as a new filter for the Stand model.  Edited the serializer to be more appropriate.  Added an API url for stands as well as changed the one for centers to be center not green (the name of the app).