Fixed add_point template
Fixed the link to the correct open street maps layer on my add_point template to be used in the POST request.  Added multiple stand objects to test this.  Also added js functions to produce template.  First one displays all the stand models on the map.  Second function displays the address as well as description of items they sell on the marker.