Changed nav bar
Added JS to center_map template in order to attempt to get GeoJson data to interact with js functions (still not working).  Changed spacing and content of the nav bar in base template.  added "paper" to the serializer for the produce stands in order to serve up that field from the model.