Fixed Produce Form
Fixed produce form to be a choice field on the form, related to name in Stand/Produce.  Created a new form for users to add information to their produce stand.  Created a new form, view, and template for the add_prod form.
Added links to base to the nav bar.